backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
sentence-transformers
faiss-cpu
openai
python-dotenv
